Moved utils classes to the utils package
