+++ even-more-fish-plugin/src/main/java/com/oheers/fish/api/EMFAPI.java
@@ -1,7 +1,7 @@
 package com.oheers.fish.api;
 
 import com.oheers.fish.EvenMoreFish;
-import com.oheers.fish.FishUtils;
+import com.oheers.fish.utils.FishUtils;
 import com.oheers.fish.exceptions.InvalidFishException;
 import com.oheers.fish.fishing.items.Fish;
 import com.oheers.fish.fishing.items.Rarity;
