+++ even-more-fish-plugin/src/main/java/com/oheers/fish/baits/BaitApplicationListener.java
@@ -1,7 +1,7 @@
 package com.oheers.fish.baits;
 
 import com.oheers.fish.EvenMoreFish;
-import com.oheers.fish.NbtUtils;
+import com.oheers.fish.utils.NbtUtils;
 import com.oheers.fish.config.messages.ConfigMessage;
 import com.oheers.fish.config.messages.Message;
 import com.oheers.fish.exceptions.MaxBaitReachedException;
