+++ even-more-fish-plugin/src/main/java/com/oheers/fish/competition/Bar.java
@@ -1,7 +1,7 @@
 package com.oheers.fish.competition;
 
 import com.oheers.fish.EvenMoreFish;
-import com.oheers.fish.FishUtils;
+import com.oheers.fish.utils.FishUtils;
 import com.oheers.fish.config.messages.ConfigMessage;
 import com.oheers.fish.config.messages.Message;
 import org.bukkit.Bukkit;
