+++ even-more-fish-plugin/src/main/java/com/oheers/fish/competition/reward/Reward.java
@@ -1,7 +1,7 @@
 package com.oheers.fish.competition.reward;
 
 import com.oheers.fish.EvenMoreFish;
-import com.oheers.fish.FishUtils;
+import com.oheers.fish.utils.FishUtils;
 import com.oheers.fish.api.EMFRewardEvent;
 import com.oheers.fish.config.messages.OldMessage;
 import me.clip.placeholderapi.PlaceholderAPI;
