+++ even-more-fish-plugin/src/main/java/com/oheers/fish/events/FishEatEvent.java
@@ -1,6 +1,6 @@
 package com.oheers.fish.events;
 
-import com.oheers.fish.FishUtils;
+import com.oheers.fish.utils.FishUtils;
 import com.oheers.fish.competition.reward.Reward;
 import com.oheers.fish.fishing.items.Fish;
 import org.bukkit.event.EventHandler;
