+++ even-more-fish-plugin/src/main/java/com/oheers/fish/fishing/items/Rarity.java
@@ -1,6 +1,6 @@
 package com.oheers.fish.fishing.items;
 
-import com.oheers.fish.FishUtils;
+import com.oheers.fish.utils.FishUtils;
 import com.oheers.fish.requirements.Requirement;
 
 import java.util.ArrayList;
