+++ even-more-fish-plugin/src/main/java/com/oheers/fish/requirements/Region.java
@@ -1,6 +1,6 @@
 package com.oheers.fish.requirements;
 
-import com.oheers.fish.FishUtils;
+import com.oheers.fish.utils.FishUtils;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.jetbrains.annotations.NotNull;
