+++ even-more-fish-plugin/src/main/java/com/oheers/fish/utils/FishUtils.java
@@ -1,7 +1,8 @@
-package com.oheers.fish;
+package com.oheers.fish.utils;
 
 import br.net.fabiozumbi12.RedProtect.Bukkit.RedProtect;
 import br.net.fabiozumbi12.RedProtect.Bukkit.Region;
+import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.config.messages.ConfigMessage;
 import com.oheers.fish.config.messages.Message;
 import com.oheers.fish.exceptions.InvalidFishException;
