+++ even-more-fish-plugin/src/main/java/com/oheers/fish/utils/ItemFactory.java
@@ -1,7 +1,6 @@
 package com.oheers.fish.utils;
 
 import com.oheers.fish.EvenMoreFish;
-import com.oheers.fish.FishUtils;
 import com.oheers.fish.config.messages.Message;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
